requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
atomicwrites==1.3.0
attrs==19.1.0
Automat==0.7.0
certifi==2018.8.24
cffi==1.11.5
click==6.7
constantly==15.1.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.28.5
Flask==1.0.2
Flask-Sijax==0.4.1
future==0.17.1
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.7
incremental==17.5.0
isort==4.3.4
itsdangerous==0.24
Jinja2==2.10
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==1.0
matplotlib==3.0.0
mccabe==0.6.1
mkl-fft==1.0.6
more-itertools==4.3.0
nltk==3.3
numpy==1.15.2
olefile==0.46
opencv-python-headless==4.1.0.25
parsel==1.5.0
pathlib2==2.3.2
Pillow==5.2.0
pluggy==0.11.0
py==1.6.0
pyasn1==0.4.5
pyasn1-modules==0.2.2
pycparser==2.19
PyDispatcher==2.0.5
pylint==2.1.1
pyOpenSSL==18.0.0
pyparsing==2.4.0
pytesseract==0.2.6
pytest==3.8.1
pytest-runner==4.4
python-dateutil==2.7.3
pytz==2019.1
queuelib==1.5.0
Scrapy==1.5.1
service-identity==17.0.0
Sijax==0.3.2
six==1.11.0
tornado==5.1.1
Twisted==17.5.0
typed-ast==1.1.0
w3lib==1.19.0
Werkzeug==0.15.2
wrapt==1.10.11
zope.interface==4.5.0
